Add files via upload
deleted duplicates from the rolling total. due to the nature of the table, I could not group by date in SQL, thus deleting duplicates in Excel saves a lot of headache in tableau/PowerBI because they will sum the values in the rolling total if you do not delete the duplicate values.